requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests
beautifulsoup4
azure-functions
azure-storage-blob
pandas
playwright==1.55.0
playwright-stealth==2.0.0
